cross-match Joe Hennawi's targets with SEQUELS results
git-svn-id: svn+ssh://astrometry.net/svn/trunk/projects/tractor@23460 03a3cea6-2c03-0410-ac9b-9271b1c66c29

+++ sequels.py
@@ -812,7 +812,7 @@ def main():
                     R2 = (T.x - cx)**2 + (T.y - cy)**2
                     J = (R2 < P.R2[I])
                     I = I[J]
-                    P.R2[I] = R2[J]
+                    P.R2[I] = R2[J].astype(np.float32)
                     #print len(I), 'are closest'
                     T.cut(J)
                 print '  ', len(T), 'from', coadd
